refactor: expose modules at top level
